Created InputWithButton component
This component is used twice on the Home screen, and consists of a View wrapping a Button on the left, and TextInput on the right. The Button's text shows a currency, and clicking it should enable the user to pick a new currency. The TextInput is to enter a numeric amount of the selected currency. The second component is passed an editable=false prop, because the user shouldnt be able to enter text there as it is view only