Changed Home screen layout with keyboard opened
The KeyboardAvoidingView component is provided by React Native and allows wrapped components to position differently when the keyboard is open. The goal was to have the two currency InputWithButton components (at the very least) still displaying with the keyboard open. The KeyboardAvoidingView component offers a 'behavior' prop with a few different options - the 'padding' option seems to work best in this case for IOS & Android. https://facebook.github.io/react-native/docs/keyboardavoidingview.html