Using custom Provider/HOC to easily display alerts
Using the 'react-native-dropdownalert' library to display alert messages to user's of the application. To give the whole applicaiton access to the alerts, a custom Provider wraps the React Native application at its entry point (just like the Redux Provider (think context)). Now in any component that needs to call the alerts (error thrown, message to display etc) you can just wrap that component with the 'connectAlert' HOC that we created (think react-redux connect) and then call the API method alertWithType that will be passed to the wrapped components props