Setup Redux Saga to listen for a few events
We should check if an API call is needed for three action types: GET_INITIAL_CONVERSION (called on initial componentDidMount), SWAP_CURRENCY, and CHANGE_BASE_CURRENCY. Redux Saga can listen to these events by listening to the actions that are emitted when these events take place. In the rootSaga generator function, we are calling yield and takeEvery three times (once for each of the target events). takeEvery listens for an 'pattern' (the event), then calls a function (the second parameter). It also implicitly passes the action along as a param to the called function.